fixed move and paste
