+++ assets/dev/js/editor/document/hooks/data/document/elements/move/link-in-link-restriction-move.js
@@ -1,12 +1,24 @@
-import LinkInLinkRestrictionBase from '../base/link-in-link-restriction-base';
+import Dependency from 'elementor-api/modules/hooks/data/dependency';
 
-export class LinkInLinkMove extends LinkInLinkRestrictionBase {
+export class LinkInLinkMove extends Dependency {
 	getCommand() {
-		return 'document/elements/move';
+		return 'document/elements/paste';
 	}
 
 	getId() {
 		return 'link-in-link-restriction-move';
+	}
+
+	apply( args ) {
+		const { containers = [ args.container ], target } = args;
+		const sourceElements = containers;
+		const targetElement = target;
+
+		const hasChildWithAnchor = sourceElements.some( ( src ) =>
+			elementor.helpers.findChildWithAnchor( src.view.el ),
+		);
+
+		return ! ( hasChildWithAnchor && elementor.helpers.findParentWithAnchor( targetElement.view.el ) );
 	}
 }
