+++ assets/dev/js/editor/document/hooks/data/document/elements/paste/link-in-link-restriction-paste.js
@@ -1,12 +1,32 @@
-import LinkInLinkRestrictionBase from '../base/link-in-link-restriction-base';
+import Dependency from 'elementor-api/modules/hooks/data/dependency';
 
-export class LinkInLinkPaste extends LinkInLinkRestrictionBase {
+export class LinkInLinkPaste extends Dependency {
 	getCommand() {
-		return 'document/elements/move';
+		return 'document/elements/paste';
 	}
 
 	getId() {
 		return 'link-in-link-restriction-paste';
+	}
+
+	apply( args ) {
+		const { containers = [ args.container ], storageKey } = args;
+		const targetElements = containers;
+		const data = elementorCommon.storage.get( storageKey );
+		const sourceElements = data.clipboard.elements;
+
+		const hasChildWithAnchor = sourceElements.some( ( src ) =>
+			src.htmlCache.indexOf( '<a' ) !== -1,
+		);
+
+		if ( ! hasChildWithAnchor ) {
+			return true;
+		}
+
+		const hasParentWithAnchor = targetElements.some( ( target ) =>
+			elementor.helpers.findParentWithAnchor( target.view.el ),
+		);
+		return ! hasParentWithAnchor;
 	}
 }
